cv-server/requirements.txt
opencv-python
numpy
pygame
pyserial
insightface
onnxruntime
python-telegram-bot==20.3
